SAN-3962; Better IPv6 Handling
The resolver now works as such:

1. Refuse all non-internal domain names (`rcode = Refused`)
2. Give an empty response for internal domains for IPv6 (`rcode = NoError`)
3. Attempt to resolve internal domains for IPv4

We had steps 2 and 3 in the wrong order and it was causing issues with external resolutions. Basically we should return `NoError` on a `AAAA` if and only if we can handle the resolution of the `A` record. We can never handle the resolution for external domain names, and thus we needed to be returning a `Refused` code to indicate as such.